# Conflicts:
#	src/main/java/com/epam/admissions_committee/controller/FacultyController.java
#	src/main/java/com/epam/admissions_committee/controller/StatementController.java
#	src/main/java/com/epam/admissions_committee/controller/UserFacultyController.java
#	src/main/java/com/epam/admissions_committee/controller/dto/FacultyDto.java
#	src/main/java/com/epam/admissions_committee/model/Applicant.java
#	src/main/java/com/epam/admissions_committee/model/Faculty.java
#	src/main/java/com/epam/admissions_committee/model/FacultyTranslate.java
#	src/main/java/com/epam/admissions_committee/model/Statement.java
#	src/main/java/com/epam/admissions_committee/model/User.java
#	src/main/java/com/epam/admissions_committee/repository/FacultyRepository.java
#	src/main/java/com/epam/admissions_committee/service/impl/FacultyServiceImpl.java
#	src/main/java/com/epam/admissions_committee/service/impl/StatementServiceImpl.java
#	src/main/java/com/epam/admissions_committee/service/impl/UserMappingServiceImpl.java
#	src/main/java/com/epam/admissions_committee/service/impl/UserServiceImpl.java

+++ src/main/java/com/epam/admissions_committee/model/Applicant.java
@@ -7,5 +7,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Applicant {
     private User applicantId;
-    private int mark;
+    private double mark;
 }
